refactor: remove unused services and implement reward event producer
- Deleted EpochService and ScoringService as they are no longer needed.
- Removed LpTrackingService due to redundancy.
- Introduced RewardEventProducer for handling reward events.
- Updated GovernanceService and BlockchainWebhookService to utilize RewardEventProducer for event indexing.
- Cleaned up imports and adjusted related logic accordingly.
