+++ src/modules/vaults/phase-management/governance/governance.service.ts
@@ -1,4 +1,4 @@
-import { BlockFrostAPI } from '@blockfrost/blockfrost-js';
+﻿import { BlockFrostAPI } from '@blockfrost/blockfrost-js';
 import {
   BadRequestException,
   ForbiddenException,
@@ -42,7 +42,7 @@ import { Vote } from '@/database/vote.entity';
 import { DexHunterPricingService } from '@/modules/dexhunter/dexhunter-pricing.service';
 import { DexHunterService } from '@/modules/dexhunter/dexhunter.service';
 import { SystemSettingsService } from '@/modules/globals/system-settings/system-settings.service';
-import { ActivityEventService } from '@/modules/rewards/services/activity-event.service';
+import { RewardEventProducer } from '@/modules/rewards/services/reward-event-producer.service';
 import { TaptoolsService } from '@/modules/taptools/taptools.service';
 import { GetAssetsToListRes } from '@/modules/vaults/phase-management/governance/dto/get-assets-to-list.res';
 import { TreasuryWalletService } from '@/modules/vaults/treasure/treasure-wallet.service';
@@ -134,8 +134,7 @@ export class GovernanceService {
     private readonly wayUpPricingService: WayUpPricingService,
     private readonly taptoolsService: TaptoolsService,
     private readonly treasuryWalletService: TreasuryWalletService,
-    private readonly activityEventService: ActivityEventService
-
+    private readonly rewardEventProducer: RewardEventProducer
   ) {
     this.isMainnet = this.configService.get<string>('CARDANO_NETWORK') === 'mainnet';
     this.poolAddress = this.configService.get<string>('POOL_ADDRESS');
@@ -345,7 +344,7 @@ export class GovernanceService {
             if (activeLpSupply === BigInt(0)) continue;
 
             for (const user of realLpHolders) {
-              // Integer division: (userLp * lockedVaultTokens) / totalLpSupply — no float rounding
+              // Integer division: (userLp * lockedVaultTokens) / totalLpSupply вЂ” no float rounding
               const underlyingTokens = (user.lpBalance * lockedVaultTokensRaw) / activeLpSupply;
               addressBalances[user.address] = (addressBalances[user.address] ?? BigInt(0)) + underlyingTokens;
             }
@@ -356,7 +355,7 @@ export class GovernanceService {
         }
 
         // Remove DEX pool contract addresses from the snapshot.
-        // Each pool contract holds VT equal to its lockedVaultTokensRaw — these tokens are
+        // Each pool contract holds VT equal to its lockedVaultTokensRaw вЂ” these tokens are
         // already redistributed proportionally to LP holders above, so keeping the pool
         // address would double-count that VT in voting power.
         for (const [address, balance] of Object.entries(addressBalances)) {
@@ -537,7 +536,7 @@ export class GovernanceService {
             }
           }
 
-          // Check BUY and OFFER assets — they are mutually exclusive for the same NFT
+          // Check BUY and OFFER assets вЂ” they are mutually exclusive for the same NFT
           // (cannot BUY if an active OFFER exists for it, and vice versa)
           const allRequestedExternalIds = [...requestedBuyAssetIds, ...requestedOfferAssetIds];
 
@@ -724,7 +723,7 @@ export class GovernanceService {
         // Validate all assets exist and handle market-specific validation
         await Promise.all(
           actions.map(async action => {
-            // BUY actions target external NFTs (not in our DB) — skip DB lookup, validate via WayUp API below
+            // BUY actions target external NFTs (not in our DB) вЂ” skip DB lookup, validate via WayUp API below
             if (action.exec === ExecType.BUY) {
               if (market === 'WayUp') {
                 const policyId = action.assetId.length >= 56 ? action.assetId.slice(0, 56) : action.assetId;
@@ -753,7 +752,7 @@ export class GovernanceService {
 
                 if (!exactAsset?.listing) {
                   throw new BadRequestException(
-                    `NFT ${displayName} is not available for purchase — no active listing found on WayUp.`
+                    `NFT ${displayName} is not available for purchase вЂ” no active listing found on WayUp.`
                   );
                 }
 
@@ -779,7 +778,7 @@ export class GovernanceService {
               return;
             }
 
-            // OFFER actions target external NFTs — validate policy whitelist and NFT existence via WayUp API
+            // OFFER actions target external NFTs вЂ” validate policy whitelist and NFT existence via WayUp API
             if (action.exec === ExecType.OFFER) {
               if (market === 'WayUp') {
                 const policyId = action.assetId.length >= 56 ? action.assetId.slice(0, 56) : action.assetId;
@@ -801,7 +800,7 @@ export class GovernanceService {
                 }
 
                 // Query WayUp to confirm the NFT exists and resolve hex assetName for nftSnapshot
-                // Unlike BUY, the NFT doesn't need to be listed — we can offer on any NFT in the collection
+                // Unlike BUY, the NFT doesn't need to be listed вЂ” we can offer on any NFT in the collection
                 const collectionResponse = await this.wayUpPricingService.getCollectionAssets({
                   policyId,
                   term: action.assetName,
@@ -1290,7 +1289,7 @@ export class GovernanceService {
     });
 
     // Index reward event for proposal creation
-    this.activityEventService.indexEvent({
+    this.rewardEventProducer.indexEvent({
       walletAddress: user.address,
       vaultId: vault.id,
       eventType: RewardActivityType.GOVERNANCE_PROPOSAL,
@@ -1780,7 +1779,7 @@ export class GovernanceService {
     await this.voteRepository.save(vote);
 
     // Index reward event for governance vote
-    this.activityEventService.indexEvent({
+    this.rewardEventProducer.indexEvent({
       walletAddress: voteReq.voterAddress,
       vaultId: proposal.vaultId,
       eventType: RewardActivityType.GOVERNANCE_VOTE,
@@ -1968,7 +1967,7 @@ export class GovernanceService {
     });
 
     // Index reward event for proposal creation (paid path)
-    this.activityEventService.indexEvent({
+    this.rewardEventProducer.indexEvent({
       walletAddress: proposal.creator.address,
       vaultId: proposal.vault.id,
       eventType: RewardActivityType.GOVERNANCE_PROPOSAL,
@@ -2800,7 +2799,7 @@ export class GovernanceService {
    * Converts a human-readable float amount (as returned by TapTools) to raw on-chain bigint units.
    * Avoids floating-point drift by parsing the fixed-decimal string representation directly.
    *
-   * Example: floatToRawBigInt(522963.34, 6) → 522963340000n
+   * Example: floatToRawBigInt(522963.34, 6) в†’ 522963340000n
    */
   private floatToRawBigInt(value: number, decimals: number): bigint {
     const fixed = value.toFixed(decimals); // e.g. "522963.340000"
