+++ src/modules/vaults/processing-tx/onchain/blockchain-webhook.service.ts
@@ -1,4 +1,4 @@
-import { verifyWebhookSignature, SignatureVerificationError } from '@blockfrost/blockfrost-js';
+﻿import { verifyWebhookSignature, SignatureVerificationError } from '@blockfrost/blockfrost-js';
 import { Injectable, Logger, UnauthorizedException } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -12,7 +12,7 @@ import { OnchainTransactionStatus } from './types/transaction-status.enum';
 import { Claim } from '@/database/claim.entity';
 import { User } from '@/database/user.entity';
 import { Vault } from '@/database/vault.entity';
-import { ActivityEventService } from '@/modules/rewards/services/activity-event.service';
+import { RewardEventProducer } from '@/modules/rewards/services/reward-event-producer.service';
 import { AssetsService } from '@/modules/vaults/assets/assets.service';
 import { ClaimStatus } from '@/types/claim.types';
 import { RewardActivityType } from '@/types/rewards.types';
@@ -37,7 +37,7 @@ export class BlockchainWebhookService {
     private readonly transactionsService: TransactionsService,
     private readonly configService: ConfigService,
     private readonly assetsService: AssetsService,
-    private readonly activityEventService: ActivityEventService,
+    private readonly rewardEventProducer: RewardEventProducer,
     @InjectRepository(Claim)
     private readonly claimRepository: Repository<Claim>,
     @InjectRepository(User)
@@ -260,7 +260,7 @@ export class BlockchainWebhookService {
         const assets = transaction.metadata as any[];
         const units = Array.isArray(assets) ? assets.reduce((sum, a) => sum + (Number(a.quantity) || 1), 0) : 1;
 
-        await this.activityEventService.indexEvent({
+        await this.rewardEventProducer.indexEvent({
           walletAddress: user.address,
           vaultId: transaction.vault_id,
           eventType: isExpansion
@@ -274,7 +274,7 @@ export class BlockchainWebhookService {
         // Acquire: use ADA amount as units (in lovelace)
         const units = transaction.amount || 1;
 
-        await this.activityEventService.indexEvent({
+        await this.rewardEventProducer.indexEvent({
           walletAddress: user.address,
           vaultId: transaction.vault_id,
           eventType: RewardActivityType.TOKEN_ACQUIRE,
@@ -287,7 +287,7 @@ export class BlockchainWebhookService {
         });
 
         // Also emit acquire phase purchase bonus
-        await this.activityEventService.indexEvent({
+        await this.rewardEventProducer.indexEvent({
           walletAddress: user.address,
           vaultId: transaction.vault_id,
           eventType: RewardActivityType.ACQUIRE_PHASE_PURCHASE,
